// types/database.types.ts
export type Json = string | number | boolean | null | { [key: string]: Json | undefined } | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: '13.0.5';
  };
  public: {
    Tables: {
// ...
      favorites: {
        Row: {
          created_at: string | null;
          favorited: boolean | null;
          id: string;
          post_id: string | null;
          user_id: string | null;
        };
        Insert: {
          created_at?: string | null;
          favorited?: boolean | null;
          id?: string;
          post_id?: string | null;
          user_id?: string | null;
        };
        Update: {
          created_at?: string | null;
          favorited?: boolean | null;
          id?: string;
          post_id?: string | null;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: 'favorites_post_id_fkey';
            columns: ['post_id'];
            isOneToOne: false;
            referencedRelation: 'posts';
            referencedColumns: ['id'];
          },
          {
            foreignKeyName: 'favorites_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      posts: {
        Row: {
          availability: boolean | null;
          beds: string | null;
          created_at: string | null;
          description: string | null;
          filters: Json | null;
          id: string;
          image: string | null;
          latitude: number | null;
          location: string | null;
          longitude: number | null;
          price_per_night: number | null;
          title: string;
          user_id: string | null;
        };
        Insert: {
          availability?: boolean | null;
          beds?: string | null;
          created_at?: string | null;
          description?: string | null;
          filters?: Json | null;
          id?: string;
          image?: string | null;
          latitude?: number | null;
          location?: string | null;
          longitude?: number | null;
          price_per_night?: number | null;
          title: string;
          user_id?: string | null;
        };
        Update: {
          availability?: boolean | null;
          beds?: string | null;
          created_at?: string | null;
          description?: string | null;
          filters?: Json | null;
          id?: string;
          image?: string | null;
          latitude?: number | null;
          location?: string | null;
          longitude?: number | null;
          price_per_night?: number | null;
          title?: string;
          user_id?: string | null;
        };
        Relationships: [
          {
            foreignKeyName: 'posts_user_id_fkey';
            columns: ['user_id'];
            isOneToOne: false;
            referencedRelation: 'users';
            referencedColumns: ['id'];
          },
        ];
      };
// ...
      users: {
        Row: {
          account_type: string | null;
          avatar: string | null;
          contact: number | null;
          created_at: string | null;
          email: string;
          firstname: string | null;
          id: string;
          landlord_proof_id: string | null;
          lastname: string | null;
          online: boolean | null;
          school: string | null;
          student_id: number | null;
          student_proof_id: string | null;
          suspended: string | null;
          username: string | null;
          verified_at: string | null;
        };
        Insert: {
          account_type?: string | null;
          avatar?: string | null;
          contact?: number | null;
          created_at?: string | null;
          email: string;
          firstname?: string | null;
          id: string;
          landlord_proof_id?: string | null;
          lastname?: string | null;
          online?: boolean | null;
          school?: string | null;
          student_id?: number | null;
          student_proof_id?: string | null;
          suspended?: string | null;
          username?: string | null;
          verified_at?: string | null;
        };
        Update: {
          account_type?: string | null;
          avatar?: string | null;
          contact?: number | null;
          created_at?: string | null;
          email?: string;
          firstname?: string | null;
          id?: string;
          landlord_proof_id?: string | null;
          lastname?: string | null;
          online?: boolean | null;
          school?: string | null;
          student_id?: number | null;
          student_proof_id?: string | null;
          suspended?: string | null;
          username?: string | null;
          verified_at?: string | null;
        };
        Relationships: [];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      requesting_user_id: { Args: never; Returns: string };
    };
    Enums: {
      [_ in never]: never;
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};